[김지예] sprint4

+++ discordit/src/main/java/com/sprint/mission/discodeit/service/basic/BasicUserService.java
@@ -1,10 +1,11 @@
 package com.sprint.mission.discodeit.service.basic;
 
+import com.sprint.mission.discodeit.dto.user.request.*;
 import com.sprint.mission.discodeit.dto.user.response.UserResponse;
 import com.sprint.mission.discodeit.entity.User;
-import com.sprint.mission.discodeit.dto.user.request.UserCreateRequest;
-import com.sprint.mission.discodeit.dto.user.request.UserUpdateRequest;
 import com.sprint.mission.discodeit.enums.OnlineStatus;
+import com.sprint.mission.discodeit.common.exceptions.binaryContent.BinaryContentNotFoundException;
+import com.sprint.mission.discodeit.common.exceptions.user.UserNotFoundException;
 import com.sprint.mission.discodeit.repository.BinaryContentRepository;
 import com.sprint.mission.discodeit.repository.ReadStatusRepository;
 import com.sprint.mission.discodeit.repository.UserRepository;
@@ -26,13 +27,9 @@ public class BasicUserService implements UserService {
     private final ReadStatusRepository readStatusRepository;
 
     @Override
-    public UserResponse getByUserId(String userId) {
-        return UserResponse.toDto(userRepository.findByUserId(userId));
-    }
-
-    @Override
-    public UserResponse getById(UUID uuid) {
-        return UserResponse.toDto(userRepository.findById(uuid));
+    public UserResponse get(UUID id) {
+        return UserResponse.toDto(userRepository.find(id)
+                .orElseThrow(() -> new UserNotFoundException(id)));
     }
 
 
@@ -53,17 +50,21 @@ public class BasicUserService implements UserService {
 
 
     @Override
-    public void signIn(UserCreateRequest dto) {
+    public UserResponse signIn(UserCreateRequest dto) {
         User user = new User(dto.id(), dto.passwd(), dto.email(), dto.displayName());
-        if (dto.profileImage() != null) {
-            user.setProfileImage(dto.profileImage());
+        if (dto.profileImageId() != null) {
+            user.setProfileImage(binaryContentRepository.findById(dto.profileImageId())
+                    .orElseThrow(() -> new BinaryContentNotFoundException(dto.profileImageId())));
         }
         userRepository.save(user);
+        return UserResponse.toDto(user);
     }
 
+    // TODO: Auth만들기
     @Override
-    public UserResponse login(String id, String passwd) {
-        User user = userRepository.findByUserId(id);
+    public UserResponse login(String userId, String passwd) {
+        User user = userRepository.findByUserId(userId)
+                .orElseThrow(() -> new UserNotFoundException(userId));
         if (!user.getPasswd().equals(passwd))
             throw new IllegalArgumentException("아이디와 비밀번호가 일치하지 않습니다.");
 
@@ -73,8 +74,9 @@ public class BasicUserService implements UserService {
     }
 
     @Override
-    public void update(UserUpdateRequest dto) {
-        User user = userRepository.findByUserId(dto.userId());
+    public UserResponse update(UserUpdateRequest dto) {
+        User user = userRepository.find(dto.id())
+                .orElseThrow(() -> new UserNotFoundException(dto.id()));
         if (dto.passwd() != null) {
             user.setPasswd(dto.passwd());
         }
@@ -90,26 +92,24 @@ public class BasicUserService implements UserService {
         if (dto.onlineStatus() != null) {
             user.setOnlineStatus(dto.onlineStatus());
         }
-        if (dto.profileImage() != null) {
-            user.setProfileImage(dto.profileImage());
+        if (dto.profileImageId() != null) {
+            user.setProfileImage(binaryContentRepository.findById(dto.profileImageId())
+                    .orElseThrow(() -> new BinaryContentNotFoundException(dto.profileImageId())));
         }
 
         userRepository.update(user);
+
+        return UserResponse.toDto(user);
     }
 
     @Override
-    public void deleteByUserId(String id){
-        User user = userRepository.findByUserId(id);
+    public void delete(UserDeleteRequest dto) {
+        User user = userRepository.find(dto.id())
+                .orElseThrow(() -> new UserNotFoundException(dto.id()));
         userRepository.delete(user);
         readStatusRepository.deleteAllByUser(user);
-        binaryContentRepository.delete(user.getProfileImage());
-    }
-
-    @Override
-    public void deleteById(UUID uuid) {
-        User user = userRepository.findById(uuid);
-        userRepository.delete(user);
-        readStatusRepository.deleteAllByUser(user);
-        binaryContentRepository.delete(user.getProfileImage());
+        if (user.getProfileImage() != null) {
+            binaryContentRepository.delete(user.getProfileImage());
+        }
     }
 }
